+++ com.unity.netcode.gameobjects/Runtime/Transports/UTP/UnityTransport.cs
@@ -15,6 +15,9 @@ using Unity.Collections;
 using Unity.Networking.Transport;
 using Unity.Networking.Transport.Relay;
 using Unity.Networking.Transport.Utilities;
+#if UTP_TRANSPORT_2_0_ABOVE
+using Unity.Networking.Transport.TLS;
+#endif
 
 #if !UTP_TRANSPORT_2_0_ABOVE
 using NetworkEndpoint = Unity.Networking.Transport.NetworkEndPoint;
@@ -164,14 +167,26 @@ namespace Unity.Netcode.Transports.UTP
         private ProtocolType m_ProtocolType;
 
 #if UTP_TRANSPORT_2_0_ABOVE
-        [Tooltip("Whether or not to use WebSockets as Network Interface")]
+        [Tooltip("Per default the client/server will communicate over UDP. Set to true to communicate with WebSocket.")]
         [SerializeField]
         private bool m_UseWebSockets = false;
 
         public bool UseWebSockets
         {
-            set => m_UseWebSockets = value;
             get => m_UseWebSockets;
+            set => m_UseWebSockets = value;
+        }
+
+        /// <summary>
+        /// Per default the client/server communication will not be encrypted. Select true to enable DTLS for UDP and TLS for Websocket.
+        /// </summary>
+        [Tooltip("Per default the client/server communication will not be encrypted. Select true to enable DTLS for UDP and TLS for Websocket.")]
+        [SerializeField]
+        private bool m_UseEncryption = false;
+        public bool UseEncryption
+        {
+            get => m_UseEncryption;
+            set => m_UseEncryption = value;
         }
 #endif
 
@@ -296,10 +311,12 @@ namespace Unity.Netcode.Transports.UTP
 
             private static NetworkEndpoint ParseNetworkEndpoint(string ip, ushort port)
             {
-                if (!NetworkEndpoint.TryParse(ip, port, out var endpoint))
+                NetworkEndpoint endpoint = default;
+
+                if (!NetworkEndpoint.TryParse(ip, port, out endpoint, NetworkFamily.Ipv4) &&
+                    !NetworkEndpoint.TryParse(ip, port, out endpoint, NetworkFamily.Ipv6))
                 {
                     Debug.LogError($"Invalid network endpoint: {ip}:{port}.");
-                    return default;
                 }
 
                 return endpoint;
@@ -491,7 +508,8 @@ namespace Unity.Netcode.Transports.UTP
 
             InitDriver();
 
-            int result = m_Driver.Bind(NetworkEndpoint.AnyIpv4);
+            var bindEndpoint = serverEndpoint.Family == NetworkFamily.Ipv6 ? NetworkEndpoint.AnyIpv6 : NetworkEndpoint.AnyIpv4;
+            int result = m_Driver.Bind(bindEndpoint);
             if (result != 0)
             {
                 Debug.LogError("Client failed to bind");
@@ -599,7 +617,7 @@ namespace Unity.Netcode.Transports.UTP
                 hostConnectionData = connectionData;
             }
 
-            m_RelayServerData = new RelayServerData(ref serverEndpoint, 1, ref allocationId, ref connectionData, ref hostConnectionData, ref key, isSecure);
+            m_RelayServerData = new RelayServerData(ref serverEndpoint, 0, ref allocationId, ref connectionData, ref hostConnectionData, ref key, isSecure);
 
             SetProtocol(ProtocolType.RelayUnityTransport);
         }
@@ -632,7 +650,7 @@ namespace Unity.Netcode.Transports.UTP
         /// <summary>
         /// Sets IP and Port information. This will be ignored if using the Unity Relay and you should call <see cref="SetRelayServerData"/>
         /// </summary>
-        /// <param name="ipv4Address">The remote IP address</param>
+        /// <param name="ipv4Address">The remote IP address (despite the name, can be an IPv6 address)</param>
         /// <param name="port">The remote port</param>
         /// <param name="listenAddress">The local listen address</param>
         public void SetConnectionData(string ipv4Address, ushort port, string listenAddress = null)
@@ -1399,6 +1417,36 @@ namespace Unity.Netcode.Transports.UTP
         }
 #endif
 
+        private FixedString4096Bytes m_ServerPrivate;
+        private FixedString4096Bytes m_ServerCertificate;
+
+        private FixedString512Bytes m_ServerCommonName;
+        private FixedString4096Bytes m_ClientCertificate;
+
+        public void SetServerSecrets(string serverCertificate, string serverPrivateKey)
+        {
+            if (serverPrivateKey.Length > m_ServerPrivate.Capacity ||
+                serverCertificate.Length > m_ServerCertificate.Capacity)
+            {
+                throw new Exception("Secret lengths are above what Unity Transport allows.");
+            }
+
+            m_ServerPrivate = serverPrivateKey;
+            m_ServerCertificate = serverCertificate;
+        }
+
+        public void SetClientSecrets(string serverCommonName, string clientCertificate = null)
+        {
+            if (serverCommonName.Length > m_ServerCommonName.Capacity ||
+                clientCertificate?.Length > m_ClientCertificate.Capacity)
+            {
+                throw new Exception("Secret lengths are above what Unity Transport allows.");
+            }
+
+            m_ServerCommonName = serverCommonName;
+            m_ClientCertificate = clientCertificate;
+        }
+
         /// <summary>
         /// Creates the internal NetworkDriver
         /// </summary>
@@ -1431,6 +1479,63 @@ namespace Unity.Netcode.Transports.UTP
                 receiveQueueCapacity: m_MaxPacketQueueSize,
 #endif
                 heartbeatTimeoutMS: transport.m_HeartbeatTimeoutMS);
+
+#if UNITY_WEBGL && !UNITY_EDITOR
+            if (NetworkManager.IsServer)
+            {
+                throw new Exception("WebGL as a server is not supported by Unity Transport, outside the Editor.");
+            }
+#endif
+
+#if UTP_TRANSPORT_2_0_ABOVE
+            if (m_UseEncryption)
+            {
+                if (m_ProtocolType == ProtocolType.RelayUnityTransport)
+                {
+                    if (m_RelayServerData.IsSecure != 0)
+                    {
+                        // log an error because we have mismatched configuration
+                        Debug.LogError("Mismatched security configuration, between Relay and local NetworkManager settings");
+                    }
+                    else
+                    {
+                        if (m_UseWebSockets)
+                        {
+                            // Todo: new code to support Relay+WSS
+                            throw new NotImplementedException();
+                        }
+                    }
+                }
+                else
+                {
+                    try
+                    {
+                        if (NetworkManager.IsServer)
+                        {
+                            if (m_ServerCertificate.Length == 0 ||
+                                m_ServerPrivate.Length == 0)
+                            {
+                                throw new Exception("In order to use encrypted communications, when hosting, you must set the server certificate and key.");
+                            }
+                            m_NetworkSettings.WithSecureServerParameters(certificate: ref m_ServerCertificate,
+                                privateKey: ref m_ServerPrivate);
+                        }
+                        else
+                        {
+                            if (m_ServerCommonName.Length == 0)
+                            {
+                                throw new Exception("In order to use encrypted communications, clients must set the server common name.");
+                            }
+                            m_NetworkSettings.WithSecureClientParameters(serverName: ref m_ServerCommonName, caCertificate: ref m_ClientCertificate);
+                        }
+                    }
+                    catch(Exception e)
+                    {
+                        Debug.LogException(e,this);
+                    }
+                }
+            }
+#endif
 
 #if UTP_TRANSPORT_2_0_ABOVE
             if (m_UseWebSockets)